requirements.txt
ml-collections==0.1.0
absl-py==0.10.0
torch>=1.7.0
torchvision